moved inverse depth measurement to vision common
