+++ localization/graph_localizer/package.xml
@@ -24,6 +24,7 @@
   <build_depend>localization_common</build_depend>
   <build_depend>localization_measurements</build_depend>
   <build_depend>msg_conversions</build_depend>
+  <build_depend>vision_common</build_depend>
   <run_depend>nodelet</run_depend>
   <run_depend>camera</run_depend>
   <run_depend>config_reader</run_depend>
@@ -34,6 +35,7 @@
   <run_depend>localization_common</run_depend>
   <run_depend>localization_measurements</run_depend>
   <run_depend>msg_conversions</run_depend>
+  <run_depend>vision_common</run_depend>
   <export>
     <nodelet plugin="${prefix}/nodelet_plugins.xml" />
   </export>
